+++ libs/mock-block-dock/src/datastore/get-default-entity-version-interval.ts
@@ -1,4 +1,4 @@
-import { EntityTemporalVersioningMetadata } from "@blockprotocol/graph";
+import { EntityTemporalVersioningMetadata } from "@blockprotocol/graph/temporal";
 
 export const getDefaultEntityVersionInterval =
   (): EntityTemporalVersioningMetadata => {
